noop sender writes to debug log only.
